+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [2.6.4] 2025-05-04 (dev)
+### Added
+- Primary/Custom menu text, such as course or category names, can now contain a pipe (|) character.
 ### Updated
 - Fixed several issues relating to Multi-language text.